Rename `ext` module to `result`
